+++ src/components/tomogram/main.tsx
@@ -20,16 +20,18 @@ import { ImageCard } from "../visualisation/image";
 import { InfoGroup } from "../visualisation/infogroup";
 import { PlotContainer } from "../visualisation/plotContainer";
 import { Motion } from "../motion/motion";
-import { useEffect, useState } from "react";
+import { Suspense, useEffect, useState } from "react";
 import { client } from "../../utils/api/client";
 import { TomogramData, BasePoint, BaseProcessingJobProps, DataConfig } from "../../schema/interfaces";
 import { CTF } from "../ctf/ctf";
 import { Scatter } from "../plots/scatter";
-import { APNGViewer } from "../visualisation/apng";
 import { setImage } from "../../utils/api/response";
 import { components } from "../../schema/main";
 import { ProcessingTitle } from "../visualisation/processingTitle";
 import { parseData } from "../../utils/generic";
+import React from "react";
+
+const APNGViewer = React.lazy(() => import("../visualisation/apng"));
 
 const tomogramConfig: DataConfig = {
   include: [
@@ -132,7 +134,9 @@ const Tomogram = ({ autoProc, procJob, status, active = false }: BaseProcessingJ
             <ModalHeader paddingBottom={0}>Movie</ModalHeader>
             <ModalCloseButton />
             <ModalBody h={{ base: "90vh", md: "60vh" }}>
-              <APNGViewer src={`tomograms/${tomogram.tomogramId}/movie`} />
+              <Suspense>
+                <APNGViewer src={`tomograms/${tomogram.tomogramId}/movie`} />
+              </Suspense>
             </ModalBody>
           </ModalContent>
         </Modal>